+++ databank/views.py
@@ -2,6 +2,8 @@ from rest_framework import mixins, viewsets
 from rest_framework.authentication import BasicAuthentication, TokenAuthentication
 from rest_framework.permissions import IsAuthenticated
 
+from main.permissions import DenyGuestUserPermission
+
 from .filter_set import FDRSIncomeFilter
 from .models import CountryOverview, FDRSIncome
 from .serializers import CountryOverviewSerializer, FDRSIncomeSerializer
@@ -11,7 +13,7 @@ class CountryOverviewViewSet(mixins.RetrieveModelMixin, viewsets.GenericViewSet)
     queryset = CountryOverview.objects.all()
     # TODO: Use global authentication class
     authentication_classes = (BasicAuthentication, TokenAuthentication)
-    permission_classes = (IsAuthenticated,)
+    permission_classes = (IsAuthenticated, DenyGuestUserPermission)
     serializer_class = CountryOverviewSerializer
     lookup_field = "country__iso__iexact"