+++ notifications/drf_views.py
@@ -8,7 +8,7 @@ from rest_framework.permissions import IsAuthenticated
 
 from deployments.models import MolnixTag
 from main.filters import CharInFilter
-from main.permissions import DenyGuestUserMutationPermission
+from main.permissions import DenyGuestUserPermission
 
 from .models import Subscription, SurgeAlert
 from .serializers import (  # UnauthenticatedSurgeAlertSerializer,
@@ -35,7 +35,13 @@ class SurgeAlertFilter(filters.FilterSet):
         help_text="Molnix_tag names, comma separated",
         widget=CSVWidget,
     )
-    status = filters.NumberFilter(field_name="status", lookup_expr="exact")
+    molnix_status = CharInFilter(
+        label="molnix status",
+        field_name="molnix_status",
+        lookup_expr="in",
+        help_text="Molnix status value, comma separated",
+        widget=CSVWidget,
+    )
 
     class Meta:
         model = SurgeAlert
@@ -43,10 +49,7 @@ class SurgeAlertFilter(filters.FilterSet):
             "created_at": ("exact", "gt", "gte", "lt", "lte"),
             "start": ("exact", "gt", "gte", "lt", "lte"),
             "end": ("exact", "gt", "gte", "lt", "lte"),
-            "is_stood_down": ("exact",),
-            "is_active": ("exact",),
             "molnix_id": ("exact", "in"),
-            "molnix_status": ("exact", "in"),
             "message": ("exact", "in"),
             "country": ("exact", "in"),
             "country__name": ("exact", "in"),
@@ -63,7 +66,7 @@ class SurgeAlertViewset(viewsets.ReadOnlyModelViewSet):
     authentication_classes = (TokenAuthentication,)
     queryset = SurgeAlert.objects.prefetch_related("molnix_tags", "molnix_tags__groups").select_related("event", "country").all()
     filterset_class = SurgeAlertFilter
-    ordering_fields = ("created_at", "atype", "category", "event", "is_stood_down", "status", "opens")
+    ordering_fields = ("created_at", "atype", "category", "event", "molnix_status", "opens")
     search_fields = (
         "operation",
         "message",
@@ -75,6 +78,10 @@ class SurgeAlertViewset(viewsets.ReadOnlyModelViewSet):
         #     return SurgeAlertSerializer
         # return UnauthenticatedSurgeAlertSerializer
         return SurgeAlertSerializer
+
+    def get_queryset(self):
+        queryset = super().get_queryset()
+        return queryset.filter(molnix_id__isnull=False).exclude(molnix_tags__name="NO_GO")
 
 
 #   def get_queryset(self):
@@ -88,7 +95,7 @@ class SurgeAlertViewset(viewsets.ReadOnlyModelViewSet):
 class SubscriptionViewset(viewsets.ModelViewSet):
     serializer_class = SubscriptionSerializer
     authentication_classes = (TokenAuthentication,)
-    permission_classes = (IsAuthenticated, DenyGuestUserMutationPermission)
+    permission_classes = (IsAuthenticated, DenyGuestUserPermission)
     search_fields = ("user__username", "rtype")  # for /docs
 
     def get_queryset(self):